+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
+
 ## Added
+- open non-editable formats on views
+- added formats for conversion docm, dot, dotx, epub, htm, html, otp, ots, ott, pot, potm, potx, pps, ppsm, ppsx,
+ pptm, rtf, xlsm, xlt, xltm, xltx
 - keep intermediate versions when editing (forcesave)
 
 ## 3.0.0